sim-lib: add simulated channel representation
Add a single representation of a simulated lightning channel which
uses our state representation to add and remove htlcs. For simplicity,
each side of the channel is represented as a separate state, with the
only interaction between the two through the changes to local balance
that happen when we settle htlcs.